move run_child_process
