+++ edge-hub/watchdog/linux/src/main.rs
@@ -2,15 +2,12 @@ mod child_process;
 
 use std::{
     io::Error,
-    process::{exit, Command, Stdio},
+    process::exit,
     sync::atomic::{AtomicBool, Ordering},
     sync::Arc,
-    thread::{sleep, spawn, JoinHandle},
-    time::Duration,
 };
 
-use child_process::ChildProcess;
-use child_process::PROCESS_POLL_INTERVAL_SECS;
+use child_process::run_child_process;
 use signal_hook::{flag::register, SIGINT, SIGTERM};
 use tracing::{error, info, subscriber, Level};
 use tracing_subscriber::fmt::Subscriber;
@@ -74,41 +71,4 @@ fn register_shutdown_listener() -> Result<Arc<AtomicBool>, Error> {
     register(SIGTERM, Arc::clone(&should_shutdown))?;
     register(SIGINT, Arc::clone(&should_shutdown))?;
     Ok(should_shutdown)
-}
-
-pub fn run_child_process(
-    name: String,
-    program: String,
-    args: String,
-    shutdown_listener: Arc<AtomicBool>,
-) -> JoinHandle<()> {
-    spawn(move || {
-        let child = match Command::new(program)
-            .arg(args)
-            .stdout(Stdio::inherit())
-            .spawn()
-        {
-            Ok(child) => {
-                info!("Launched {:?} process with pid {:?}", name, child.id());
-                child
-            }
-            Err(e) => {
-                error!("Failed to start {:?} process. {:?}", name, e);
-                exit(2);
-            }
-        };
-        let mut child_process = ChildProcess {
-            name,
-            process: child,
-        };
-
-        while child_process.is_running() && !shutdown_listener.load(Ordering::Relaxed) {
-            let poll_interval: Duration = Duration::from_secs(PROCESS_POLL_INTERVAL_SECS);
-            sleep(poll_interval);
-        }
-
-        shutdown_listener.store(true, Ordering::Relaxed); // tell the threads to shut down their child process
-
-        child_process.shutdown_if_running();
-    })
 }
